+++ api/routes/mpc/ecdsa/secret.ts
@@ -1,6 +1,7 @@
 import { Context } from "@crypto-mpc";
 import { SocketStream } from "@fastify/websocket";
 import { db } from "@lib/dev-db";
+import { ActionStatus } from "../mpc-routes";
 import { step } from "../step";
 
 export const initGenerateGenericSecret = (connection: SocketStream) => {
@@ -9,19 +10,17 @@ export const initGenerateGenericSecret = (connection: SocketStream) => {
   connection.socket.on("message", (message) => {
     if (!context) context = Context.createGenerateGenericSecretContext(2, 256);
 
-    const messageFromClient = new Uint8Array(message as ArrayBuffer);
-
-    const stepOutput = step(messageFromClient, context);
+    const stepOutput = step(message.toString(), context);
 
     if (stepOutput === true) {
       // TODO: Remove this in favor of real database
       db.shareBuf = context.getNewShare();
-      console.log("share ", db.shareBuf);
 
       connection.socket.close();
+      return;
     }
 
-    connection.socket.send(JSON.stringify(stepOutput));
+    connection.socket.send(stepOutput as string);
   });
 
   connection.socket.on("error", (err) => {
@@ -31,26 +30,37 @@ export const initGenerateGenericSecret = (connection: SocketStream) => {
 
 export const importGenericSecret = (connection: SocketStream) => {
   let context: Context;
+  let status: ActionStatus = "Init";
 
   connection.socket.on("message", (message) => {
-    
-    const messageFromClient = new Uint8Array(message as ArrayBuffer);
+    switch (status) {
+      case "Init":
+        context = Context.createImportGenericSecretContext(2, 256, message);
+        status = "Stepping";
+        connection.socket.send(JSON.stringify({ value: "Start" }));
+        break;
 
-    if (!context) context = Context.createImportGenericSecretContext(2, 256, message as ArrayBuffer);
-
-    const stepOutput = step(messageFromClient, context);
-
-    if (stepOutput === true) {
-      // TODO: Remove this in favor of real database
-      db.shareBuf = context.getNewShare();
-
-      connection.socket.close();
+      case "Stepping":
+        stepWithMessage(connection, message, context);
+        break;
     }
-
-    connection.socket.send(JSON.stringify(stepOutput));
   });
 
   connection.socket.on("error", (err) => {
     console.log("error", err);
   });
+};
+
+const stepWithMessage = (connection, message, context): void => {
+  const stepOutput = step(message.toString(), context);
+
+  if (stepOutput === true) {
+    // TODO: Remove this in favor of real database
+    db.shareBuf = context.getNewShare();
+
+    connection.socket.close();
+    return;
+  }
+
+  connection.socket.send(stepOutput);
 };
